Rename per-namespace base.rb to disambiguate
lib/{validator,submitter}/base.rb and lib/validator/auto_annotator/base.rb
all defined different top-level Base classes. Once Zeitwerk takes over
autoloading these files (next commit), three files mapping to the same
'Base' constant collide. Rename each file to match its constant
(validator_base.rb, submitter_base.rb, auto_annotator_base.rb) and
update the require_relative call sites. No behavior change.
