+++ lib/submitter/biosample_submitter.rb
@@ -1,4 +1,4 @@
-require_relative 'base'
+require_relative 'submitter_base'
 
 class BioSampleSubmitter < SubmitterBase
   BIOSAMPLE_DB_NAME = 'biosample'
