+++ lib/validator/auto_annotator/auto_annotator_base.rb
