+++ lib/validator/auto_annotator/auto_annotator_json.rb
@@ -1,5 +1,5 @@
 require 'fileutils'
-require_relative 'base'
+require_relative 'auto_annotator_base'
 
 #
 # A class for Auto-annotation. JSON file base.
