+++ lib/validator/common/xml_convertor.rb
@@ -1,4 +1,4 @@
-require_relative '../base'
+require_relative '../validator_base'
 
 #
 # XMLの変換処理を行うクラス
