+++ lib/validator/metabobank_idf_validator.rb
@@ -1,4 +1,4 @@
-require_relative 'base'
+require_relative 'validator_base'
 require_relative 'common/insdc_nullability'
 require_relative 'common/tsv_field_validator'
 require_relative 'common/file_parser'
