+++ lib/validator/trad_validator.rb
@@ -1,5 +1,5 @@
 require 'date'
-require_relative 'base'
+require_relative 'validator_base'
 require_relative 'common/date_format'
 require_relative 'common/ddbj_db_validator'
 require_relative 'common/organism_validator'
