+++ lib/validator/validator_base.rb
