+++ test/lib/validator/auto_annotator/auto_annotator_base_test.rb
@@ -1,5 +1,5 @@
 require 'test_helper'
-require 'validator/auto_annotator/base'
+require 'validator/auto_annotator/auto_annotator_base'
 
 # auto_annotationのエラー情報で元ファイルから補正後のファイルが正しく出力できるか確認
 #
